Move detect_color_card to its own module
