+++ plantcv/plantcv/transform/__init__.py
@@ -10,8 +10,7 @@ from plantcv.plantcv.transform.color_correction import quick_color_check
 from plantcv.plantcv.transform.color_correction import find_color_card
 from plantcv.plantcv.transform.color_correction import std_color_matrix
 from plantcv.plantcv.transform.color_correction import affine_color_correction
-from plantcv.plantcv.transform.color_correction import detect_color_card
-from plantcv.plantcv.transform.color_correction import _is_square
+from plantcv.plantcv.transform.detect_color_card import detect_color_card
 from plantcv.plantcv.transform.rescale import rescale
 from plantcv.plantcv.transform.rotate import rotate
 from plantcv.plantcv.transform.nonuniform_illumination import nonuniform_illumination
@@ -22,4 +21,4 @@ from plantcv.plantcv.transform.gamma_correct import gamma_correct
 __all__ = ["get_color_matrix", "get_matrix_m", "calc_transformation_matrix", "apply_transformation_matrix",
            "save_matrix", "load_matrix", "correct_color", "create_color_card_mask", "quick_color_check",
            "find_color_card", "std_color_matrix", "affine_color_correction", "rescale", "nonuniform_illumination", "resize",
-           "resize_factor", "warp", "rotate", "warp", "warp_align", "gamma_correct", "detect_color_card", "_is_square"]
+           "resize_factor", "warp", "rotate", "warp", "warp_align", "gamma_correct", "detect_color_card"]
